sui-deta-backend/main/requirements.txt
deta
requests
fastapi
uvicorn
pytz
arrow
quickchart.io